initial commit: add yaml metadata script from siafoo
http://www.siafoo.net/snippet/186

This an updated version of the script above available at siafoo. The goal is to be able to generate SQL Alchemy metadata and classes from YAML files. This would allow development shops using different languages to be able to generate their database model from a common set of yaml definitions

TODO:

- column attributes
- tests
- ?